+++ Nickel/Framework/Content/CharacterManager.cs
@@ -10,16 +10,19 @@ namespace Nickel;
 
 internal sealed class CharacterManager
 {
-	private Func<IModManifest, ILogger> LoggerProvider { get; }
-	private SpriteManager Sprites { get; }
-	private DeckManager Decks { get; }
-	private StatusManager Statuses { get; }
-	private IModManifest VanillaModManifest { get; }
-	private AfterDbInitManager<AnimationEntry> AnimationManager { get; }
-	private AfterDbInitManager<CharacterEntry> CharManager { get; }
-	private Dictionary<string, AnimationEntry> UniqueNameToAnimationEntry { get; } = [];
-	private Dictionary<string, CharacterEntry> UniqueNameToCharacterEntry { get; } = [];
-	private Dictionary<Deck, CharacterEntry> DeckToCharacterEntry { get; } = [];
+	private readonly Func<IModManifest, ILogger> LoggerProvider;
+	private readonly SpriteManager Sprites;
+	private readonly DeckManager Decks;
+	private readonly StatusManager Statuses;
+	private readonly IModManifest VanillaModManifest;
+	private readonly AfterDbInitManager<AnimationEntry> AnimationManager;
+	private readonly AfterDbInitManager<PlayableCharacterEntry> PlayableCharacterManager;
+	private readonly AfterDbInitManager<NonPlayableCharacterEntry> NonPlayableCharacterManager;
+	private readonly Dictionary<string, AnimationEntry> UniqueNameToAnimationEntry = [];
+	private readonly Dictionary<string, PlayableCharacterEntry> UniqueNameToPlayableCharacterEntry = [];
+	private readonly Dictionary<string, NonPlayableCharacterEntry> UniqueNameToNonPlayableCharacterEntry = [];
+	private readonly Dictionary<Deck, PlayableCharacterEntry> DeckToCharacterEntry = [];
+	private readonly Dictionary<string, ICharacterEntryV2> CharacterTypeToCharacterEntry = [];
 
 	public CharacterManager(
 		Func<ModLoadPhase> currentModLoadPhaseProvider,
@@ -36,7 +39,8 @@ internal sealed class CharacterManager
 		this.Statuses = statuses;
 		this.VanillaModManifest = vanillaModManifest;
 		this.AnimationManager = new(currentModLoadPhaseProvider, Inject);
-		this.CharManager = new(currentModLoadPhaseProvider, this.Inject);
+		this.PlayableCharacterManager = new(currentModLoadPhaseProvider, this.Inject);
+		this.NonPlayableCharacterManager = new(currentModLoadPhaseProvider, this.Inject);
 
 		EventsPatches.OnCrystallizedFriendEvent.Subscribe(this.OnCrystallizedFriendEvent);
 		StatePatches.OnModifyPotentialExeCards.Subscribe(this.OnModifyPotentialExeCards);
@@ -47,21 +51,152 @@ internal sealed class CharacterManager
 	internal void InjectQueuedEntries()
 	{
 		this.AnimationManager.InjectQueuedEntries();
-		this.CharManager.InjectQueuedEntries();
+		this.PlayableCharacterManager.InjectQueuedEntries();
 	}
 
 	internal void InjectLocalizations(string locale, Dictionary<string, string> localizations)
 	{
-		foreach (var entry in this.UniqueNameToCharacterEntry.Values)
+		foreach (var entry in this.UniqueNameToPlayableCharacterEntry.Values)
 			this.InjectLocalization(locale, localizations, entry);
 	}
-
-	public ICharacterAnimationEntry RegisterCharacterAnimation(IModManifest owner, string name, CharacterAnimationConfiguration configuration)
+	
+	private PlayableCharacterEntry CreateForVanilla(Deck deck)
 	{
+		var borderSprite = DB.charPanels[deck.Key()];
+		var starters = StarterDeck.starterSets[deck];
+		var neutralAnimationFrames = DB.charAnimations[Character.GetSpriteAliasIfExists(deck.Key())]["neutral"];
+		var miniAnimationFrames = DB.charAnimations[Character.GetSpriteAliasIfExists(deck.Key())]["mini"];
+		var startLocked = deck is Deck.goat or Deck.eunice or Deck.hacker or Deck.shard;
+		var missingStatusColor = DB.statuses[StatusMeta.deckToMissingStatus[deck]].color;
+		var missingStatusSprite = DB.statuses[StatusMeta.deckToMissingStatus[deck]].icon;
+		var exeCardType = deck switch
+		{
+			Deck.dizzy => typeof(ColorlessDizzySummon),
+			Deck.riggs => typeof(ColorlessRiggsSummon),
+			Deck.peri => typeof(ColorlessPeriSummon),
+			Deck.goat => typeof(ColorlessIsaacSummon),
+			Deck.eunice => typeof(ColorlessDrakeSummon),
+			Deck.hacker => typeof(ColorlessMaxSummon),
+			Deck.shard => typeof(ColorlessBooksSummon),
+			Deck.colorless => typeof(ColorlessCATSummon),
+			_ => null
+		};
+		SingleLocalizationProvider description = _ => Loc.T($"char.{deck}.desc");
+		
+		return new PlayableCharacterEntry(
+			modOwner: this.VanillaModManifest,
+			uniqueName: Enum.GetName(deck)!,
+			v1: new()
+			{
+				Deck = deck,
+				BorderSprite = borderSprite,
+				Starters = starters,
+				NeutralAnimation = new()
+				{
+					Deck = deck,
+					LoopTag = "neutral",
+					Frames = neutralAnimationFrames
+				},
+				MiniAnimation = new()
+				{
+					Deck = deck,
+					LoopTag = "mini",
+					Frames = miniAnimationFrames
+				},
+				StartLocked = startLocked,
+				MissingStatus = new()
+				{
+					Color = missingStatusColor,
+					Sprite = missingStatusSprite
+				},
+				ExeCardType = exeCardType,
+				Description = description
+			},
+			v2: new()
+			{
+				Deck = deck,
+				BorderSprite = borderSprite,
+				Starters = starters,
+				NeutralAnimation = new()
+				{
+					CharacterType = deck == Deck.colorless ? "comp" : deck.Key(),
+					LoopTag = "neutral",
+					Frames = neutralAnimationFrames
+				},
+				MiniAnimation = new()
+				{
+					CharacterType = deck == Deck.colorless ? "comp" : deck.Key(),
+					LoopTag = "mini",
+					Frames = miniAnimationFrames
+				},
+				StartLocked = startLocked,
+				MissingStatus = new()
+				{
+					Color = missingStatusColor,
+					Sprite = missingStatusSprite
+				},
+				ExeCardType = exeCardType,
+				Description = description
+			},
+			missingStatus: this.Statuses.LookupByStatus(StatusMeta.deckToMissingStatus[deck])!
+		);
+	}
+
+	private IStatusEntry RegisterMissingStatus(
+		IModManifest owner,
+		string name,
+		IDeckEntry deckEntry,
+		Color? configurationColor,
+		Spr? configurationSprite
+	)
+	{
+		var color = configurationColor ?? deckEntry.Configuration.Definition.color;
+		var sprite = configurationSprite ?? this.Sprites.RegisterSprite(owner, $"{name}::Icon", () =>
+		{
+			var questionMarkTexture = SpriteLoader.Get(Enum.Parse<Spr>("icons_questionMark"))!;
+			var data = new MGColor[questionMarkTexture.Width * questionMarkTexture.Height];
+			questionMarkTexture.GetData(data);
+
+			for (var i = 0; i < data.Length; i++)
+				data[i] = new MGColor(
+					(float)(data[i].R / 255.0 * color.r),
+					(float)(data[i].G / 255.0 * color.g),
+					(float)(data[i].B / 255.0 * color.b),
+					(float)(data[i].A / 255.0 * color.a)
+				);
+
+			var texture = new Texture2D(MG.inst.GraphicsDevice, questionMarkTexture.Width, questionMarkTexture.Height);
+			texture.SetData(data);
+			return texture;
+		}).Sprite;
+
+		return this.Statuses.RegisterStatus(owner, name, new()
+		{
+			Definition = new()
+			{
+				color = color,
+				icon = sprite,
+				isGood = false
+			},
+			Name = locale => $"{deckEntry.Configuration.Name?.Invoke(locale)} is missing", // TODO: localize
+			Description = locale => $"The next {{0}} <c={color}>{deckEntry.Configuration.Name?.Invoke(locale)}</c> cards you play do nothing.", // TODO: localize
+		});
+	}
+	
+	#region V1
+	public ICharacterAnimationEntry RegisterCharacterAnimation(IModManifest owner, string name, CharacterAnimationConfiguration v1)
+	{
+		var v2 = new CharacterAnimationConfigurationV2
+		{
+			CharacterType = v1.Deck.Key(),
+			LoopTag = v1.LoopTag,
+			Frames = v1.Frames
+		};
+		
 		var uniqueName = $"{owner.UniqueName}::{name}";
 		if (this.UniqueNameToAnimationEntry.ContainsKey(uniqueName))
 			throw new ArgumentException($"A character animation with the unique name `{uniqueName}` is already registered", nameof(name));
-		AnimationEntry entry = new(owner, uniqueName, configuration);
+		var entry = new AnimationEntry(owner, uniqueName, v1, v2);
 		this.UniqueNameToAnimationEntry[entry.UniqueName] = entry;
 		this.AnimationManager.QueueOrInject(entry);
 		return entry;
@@ -73,114 +208,198 @@ internal sealed class CharacterManager
 			return entry;
 		if (deck is not (Deck.dizzy or Deck.riggs or Deck.peri or Deck.goat or Deck.eunice or Deck.hacker or Deck.shard or Deck.colorless))
 			return null;
-
-		return new CharacterEntry(
-			modOwner: this.VanillaModManifest,
-			uniqueName: Enum.GetName(deck)!,
-			configuration: new()
-			{
-				Deck = deck,
-				BorderSprite = DB.charPanels[deck.Key()],
-				Starters = StarterDeck.starterSets[deck],
-				NeutralAnimation = new()
-				{
-					Deck = deck,
-					LoopTag = "neutral",
-					Frames = DB.charAnimations[Character.GetSpriteAliasIfExists(deck.Key())]["neutral"]
-				},
-				MiniAnimation = new()
-				{
-					Deck = deck,
-					LoopTag = "mini",
-					Frames = DB.charAnimations[Character.GetSpriteAliasIfExists(deck.Key())]["mini"]
-				},
-				StartLocked = deck is Deck.goat or Deck.eunice or Deck.hacker or Deck.shard,
-				MissingStatus = new()
-				{
-					Color = DB.statuses[StatusMeta.deckToMissingStatus[deck]].color,
-					Sprite = DB.statuses[StatusMeta.deckToMissingStatus[deck]].icon
-				},
-				ExeCardType = deck switch
-				{
-					Deck.dizzy => typeof(ColorlessDizzySummon),
-					Deck.riggs => typeof(ColorlessRiggsSummon),
-					Deck.peri => typeof(ColorlessPeriSummon),
-					Deck.goat => typeof(ColorlessIsaacSummon),
-					Deck.eunice => typeof(ColorlessDrakeSummon),
-					Deck.hacker => typeof(ColorlessMaxSummon),
-					Deck.shard => typeof(ColorlessBooksSummon),
-					Deck.colorless => typeof(ColorlessCATSummon),
-					_ => null
-				},
-				Description = _ => Loc.T($"char.{deck}.desc")
-			},
-			missingStatus: this.Statuses.LookupByStatus(StatusMeta.deckToMissingStatus[deck])!
-		);
+		return this.CreateForVanilla(deck);
 	}
 
 	public ICharacterEntry? LookupByUniqueName(string uniqueName)
-		=> this.UniqueNameToCharacterEntry.GetValueOrDefault(uniqueName);
+		=> this.UniqueNameToPlayableCharacterEntry.GetValueOrDefault(uniqueName);
 
-	public ICharacterEntry RegisterCharacter(IModManifest owner, string name, CharacterConfiguration configuration)
+	public ICharacterEntry RegisterCharacter(IModManifest owner, string name, CharacterConfiguration v1)
 	{
+		if (this.Decks.LookupByDeck(v1.Deck) is not { } deckEntry)
+			throw new ArgumentException("Invalid character `Deck`");
+		
 #pragma warning disable CS0618 // Type or member is obsolete
-		if (configuration.Starters is not null && (configuration.StarterCardTypes is not null || configuration.StarterArtifactTypes is not null))
+		if (v1.Starters is not null && (v1.StarterCardTypes is not null || v1.StarterArtifactTypes is not null))
 			throw new ArgumentException($"A character should only have `{nameof(CharacterConfiguration.Starters)}` or `{nameof(CharacterConfiguration.StarterCardTypes)}`/`{nameof(CharacterConfiguration.StarterArtifactTypes)}` defined, but not both");
+		
+		var v2 = new PlayableCharacterConfigurationV2
+		{
+			Deck = v1.Deck,
+			BorderSprite = v1.BorderSprite,
+			Starters = v1.Starters ?? new()
+			{
+				artifacts = v1.StarterArtifactTypes?.Select(t => (Artifact)Activator.CreateInstance(t)!).ToList() ?? [],
+				cards = v1.StarterCardTypes?.Select(t => (Card)Activator.CreateInstance(t)!).ToList() ?? []
+			},
+			NeutralAnimation = v1.NeutralAnimation is { } v1NeutralAnimation
+				? new()
+				{
+					CharacterType = v1.Deck.Key(),
+					LoopTag = v1NeutralAnimation.LoopTag,
+					Frames = v1NeutralAnimation.Frames
+				} : null,
+			MiniAnimation = v1.MiniAnimation is { } v1MiniAnimation
+				? new()
+				{
+					CharacterType = v1.Deck.Key(),
+					LoopTag = v1MiniAnimation.LoopTag,
+					Frames = v1MiniAnimation.Frames
+				} : null,
+			StartLocked = v1.StartLocked,
+			MissingStatus = new()
+			{
+				Color = v1.MissingStatus.Color,
+				Sprite = v1.MissingStatus.Sprite
+			},
+			ExeCardType = v1.ExeCardType,
+			Description = v1.Description
+		};
 #pragma warning restore CS0618 // Type or member is obsolete
 
 		var uniqueName = $"{owner.UniqueName}::{name}";
-		if (this.UniqueNameToCharacterEntry.ContainsKey(uniqueName))
+		if (this.UniqueNameToPlayableCharacterEntry.ContainsKey(uniqueName))
 			throw new ArgumentException($"A character with the unique name `{uniqueName}` is already registered", nameof(name));
-		var missingStatus = this.RegisterMissingStatus(owner, $"{name}::MissingStatus", configuration, configuration.MissingStatus);
-		CharacterEntry entry = new(owner, uniqueName, configuration, missingStatus);
-		this.UniqueNameToCharacterEntry[entry.UniqueName] = entry;
-		this.DeckToCharacterEntry[entry.Configuration.Deck] = entry;
-		this.CharManager.QueueOrInject(entry);
+		
+		var missingStatus = this.RegisterMissingStatus(owner, $"{name}::MissingStatus", deckEntry, v1.MissingStatus.Color, v1.MissingStatus.Sprite);
+		var entry = new PlayableCharacterEntry(owner, uniqueName, v1, v2, missingStatus);
+		this.UniqueNameToPlayableCharacterEntry[entry.UniqueName] = entry;
+		this.DeckToCharacterEntry[entry.V2.Deck] = entry;
+		this.CharacterTypeToCharacterEntry[entry.CharacterType] = entry;
+		this.PlayableCharacterManager.QueueOrInject(entry);
+		return entry;
+	}
+	#endregion
+
+	#region V2
+	public ICharacterAnimationEntryV2 RegisterCharacterAnimationV2(IModManifest owner, string name, CharacterAnimationConfigurationV2 v2)
+	{
+		var uniqueName = $"{owner.UniqueName}::{name}";
+		if (this.UniqueNameToAnimationEntry.ContainsKey(uniqueName))
+			throw new ArgumentException($"A character animation with the unique name `{uniqueName}` is already registered", nameof(name));
+		var entry = new AnimationEntry(owner, uniqueName, null, v2);
+		this.UniqueNameToAnimationEntry[entry.UniqueName] = entry;
+		this.AnimationManager.QueueOrInject(entry);
 		return entry;
 	}
 
-	private IStatusEntry RegisterMissingStatus(
-		IModManifest owner,
-		string name,
-		CharacterConfiguration characterConfiguration,
-		CharacterConfiguration.MissingStatusConfiguration configuration
-	)
+	public IPlayableCharacterEntryV2? LookupByDeckV2(Deck deck)
 	{
-		var deck = this.Decks.LookupByDeck(characterConfiguration.Deck) ?? throw new ArgumentException("Invalid character `Deck`");
-		var sprite = configuration.Sprite;
-		var color = configuration.Color ?? deck.Configuration.Definition.color;
-
-		sprite ??= this.Sprites.RegisterSprite(owner, $"{name}::Icon", () =>
-			{
-				var questionMarkTexture = SpriteLoader.Get(Enum.Parse<Spr>("icons_questionMark"))!;
-				var data = new MGColor[questionMarkTexture.Width * questionMarkTexture.Height];
-				questionMarkTexture.GetData(data);
-
-				for (var i = 0; i < data.Length; i++)
-					data[i] = new MGColor(
-						(float)(data[i].R / 255.0 * color.r),
-						(float)(data[i].G / 255.0 * color.g),
-						(float)(data[i].B / 255.0 * color.b),
-						(float)(data[i].A / 255.0 * color.a)
-					);
-
-				var texture = new Texture2D(MG.inst.GraphicsDevice, questionMarkTexture.Width, questionMarkTexture.Height);
-				texture.SetData(data);
-				return texture;
-			}).Sprite;
-
-		return this.Statuses.RegisterStatus(owner, name, new()
-		{
-			Definition = new()
-			{
-				color = color,
-				icon = sprite.Value,
-				isGood = false
-			},
-			Name = locale => $"{deck.Configuration.Name?.Invoke(locale)} is missing", // TODO: localize
-			Description = locale => $"The next {{0}} <c={color}>{deck.Configuration.Name?.Invoke(locale)}</c> cards you play do nothing.", // TODO: localize
-		});
+		if (this.DeckToCharacterEntry.TryGetValue(deck, out var entry))
+			return entry;
+		if (deck is not (Deck.dizzy or Deck.riggs or Deck.peri or Deck.goat or Deck.eunice or Deck.hacker or Deck.shard or Deck.colorless))
+			return null;
+		return this.CreateForVanilla(deck);
 	}
+
+	public ICharacterEntryV2? LookupByCharacterTypeV2(string characterType)
+	{
+		if (this.CharacterTypeToCharacterEntry.TryGetValue(characterType, out var entry))
+			return entry;
+		
+		switch (characterType)
+		{
+			case "dizzy":
+				return this.LookupByDeckV2(Deck.dizzy);
+			case "riggs":
+				return this.LookupByDeckV2(Deck.riggs);
+			case "peri":
+				return this.LookupByDeckV2(Deck.peri);
+			case "goat":
+				return this.LookupByDeckV2(Deck.goat);
+			case "eunice":
+				return this.LookupByDeckV2(Deck.eunice);
+			case "hacker":
+				return this.LookupByDeckV2(Deck.hacker);
+			case "shard":
+				return this.LookupByDeckV2(Deck.shard);
+			case "comp":
+				return this.LookupByDeckV2(Deck.colorless);
+			default:
+				if (!DB.currentLocale.strings.ContainsKey($"char.{characterType}"))
+					return null;
+				return new NonPlayableCharacterEntry(
+					this.VanillaModManifest,
+					characterType,
+					new()
+					{
+						CharacterType = characterType,
+						BorderSprite = DB.charPanels.TryGetValue(characterType, out var borderSprite) ? borderSprite : null,
+						NeutralAnimation = new()
+						{
+							CharacterType = characterType,
+							LoopTag = "neutral",
+							Frames = DB.charAnimations[Character.GetSpriteAliasIfExists(characterType)]["neutral"]
+						},
+						Name = _ => Loc.T($"char.{characterType}")
+					}
+				);
+		}
+	}
+
+	public ICharacterEntryV2? LookupByUniqueNameV2(string uniqueName)
+		=> this.UniqueNameToPlayableCharacterEntry.GetValueOrDefault(uniqueName);
+	
+	public IPlayableCharacterEntryV2 RegisterPlayableCharacterV2(IModManifest owner, string name, PlayableCharacterConfigurationV2 v2)
+	{
+		if (this.Decks.LookupByDeck(v2.Deck) is not { } deckEntry)
+			throw new ArgumentException("Invalid character `Deck`");
+		
+		var v1 = new CharacterConfiguration
+		{
+			Deck = v2.Deck,
+			BorderSprite = v2.BorderSprite,
+			Starters = v2.Starters,
+			NeutralAnimation = v2.NeutralAnimation is { } v2NeutralAnimation
+				? new()
+				{
+					Deck = v2.Deck,
+					LoopTag = v2NeutralAnimation.LoopTag,
+					Frames = v2NeutralAnimation.Frames
+				} : null,
+			MiniAnimation = v2.MiniAnimation is { } v2MiniAnimation
+				? new()
+				{
+					Deck = v2.Deck,
+					LoopTag = v2MiniAnimation.LoopTag,
+					Frames = v2MiniAnimation.Frames
+				} : null,
+			StartLocked = v2.StartLocked,
+			MissingStatus = new()
+			{
+				Color = v2.MissingStatus.Color,
+				Sprite = v2.MissingStatus.Sprite
+			},
+			ExeCardType = v2.ExeCardType,
+			Description = v2.Description
+		};
+
+		var uniqueName = $"{owner.UniqueName}::{name}";
+		if (this.UniqueNameToPlayableCharacterEntry.ContainsKey(uniqueName))
+			throw new ArgumentException($"A character with the unique name `{uniqueName}` is already registered", nameof(name));
+		
+		var missingStatus = this.RegisterMissingStatus(owner, $"{name}::MissingStatus", deckEntry, v2.MissingStatus.Color, v2.MissingStatus.Sprite);
+		var entry = new PlayableCharacterEntry(owner, uniqueName, v1, v2, missingStatus);
+		this.UniqueNameToPlayableCharacterEntry[entry.UniqueName] = entry;
+		this.DeckToCharacterEntry[entry.V2.Deck] = entry;
+		this.CharacterTypeToCharacterEntry[entry.CharacterType] = entry;
+		this.PlayableCharacterManager.QueueOrInject(entry);
+		return entry;
+	}
+	
+	public INonPlayableCharacterEntryV2 RegisterNonPlayableCharacterV2(IModManifest owner, string name, NonPlayableCharacterConfigurationV2 v2)
+	{
+		var uniqueName = $"{owner.UniqueName}::{name}";
+		if (this.UniqueNameToPlayableCharacterEntry.ContainsKey(uniqueName))
+			throw new ArgumentException($"A character with the unique name `{uniqueName}` is already registered", nameof(name));
+		
+		var entry = new NonPlayableCharacterEntry(owner, uniqueName, v2);
+		this.UniqueNameToNonPlayableCharacterEntry[entry.UniqueName] = entry;
+		this.CharacterTypeToCharacterEntry[entry.CharacterType] = entry;
+		this.NonPlayableCharacterManager.QueueOrInject(entry);
+		return entry;
+	}
+	#endregion
 
 	public bool TryGetCharacterAnimationByUniqueName(string uniqueName, [MaybeNullWhen(false)] out ICharacterAnimationEntry entry)
 	{
@@ -198,7 +417,7 @@ internal sealed class CharacterManager
 
 	public bool TryGetCharacterByUniqueName(string uniqueName, [MaybeNullWhen(false)] out ICharacterEntry entry)
 	{
-		if (this.UniqueNameToCharacterEntry.TryGetValue(uniqueName, out var typedEntry))
+		if (this.UniqueNameToPlayableCharacterEntry.TryGetValue(uniqueName, out var typedEntry))
 		{
 			entry = typedEntry;
 			return true;
@@ -212,38 +431,38 @@ internal sealed class CharacterManager
 
 	private static void Inject(AnimationEntry entry)
 	{
-		if (!DB.charAnimations.TryGetValue(entry.Configuration.Deck.Key(), out var characterAnimations))
+		if (!DB.charAnimations.TryGetValue(entry.V2.CharacterType, out var characterAnimations))
 		{
 			characterAnimations = [];
-			DB.charAnimations[entry.Configuration.Deck.Key()] = characterAnimations;
+			DB.charAnimations[entry.V2.CharacterType] = characterAnimations;
 		}
-		characterAnimations[entry.Configuration.LoopTag] = entry.Configuration.Frames.ToList();
+		characterAnimations[entry.V2.LoopTag] = entry.V2.Frames.ToList();
 	}
 
-	private void Inject(CharacterEntry entry)
+	private void Inject(PlayableCharacterEntry entry)
 	{
-		if (entry.Configuration.NeutralAnimation is { } neutralAnimationConfiguration)
+		if (entry.V2.NeutralAnimation is { } neutralAnimationConfiguration)
 		{
 			if (neutralAnimationConfiguration.LoopTag != "neutral")
 			{
 				this.LoggerProvider(entry.ModOwner).LogError($"Could not inject character {{Character}}: `{nameof(CharacterConfiguration.NeutralAnimation)}` is not tagged `neutral`.", entry.UniqueName);
 				return;
 			}
-			this.RegisterCharacterAnimation(entry.ModOwner, $"{entry.UniqueName}::neutral", neutralAnimationConfiguration);
+			this.RegisterCharacterAnimationV2(entry.ModOwner, $"{entry.UniqueName}::neutral", neutralAnimationConfiguration);
 		}
-		if (entry.Configuration.MiniAnimation is { } miniAnimationConfiguration)
+		if (entry.V2.MiniAnimation is { } miniAnimationConfiguration)
 		{
 			if (miniAnimationConfiguration.LoopTag != "mini")
 			{
 				this.LoggerProvider(entry.ModOwner).LogError($"Could not inject character {{Character}}: `{nameof(CharacterConfiguration.MiniAnimation)}` is not tagged `mini`.", entry.UniqueName);
 				return;
 			}
-			this.RegisterCharacterAnimation(entry.ModOwner, $"{entry.UniqueName}::mini", miniAnimationConfiguration);
+			this.RegisterCharacterAnimationV2(entry.ModOwner, $"{entry.UniqueName}::mini", miniAnimationConfiguration);
 		}
 
-		if (entry.Configuration.NeutralAnimation is null || entry.Configuration.MiniAnimation is null)
+		if (entry.V2.NeutralAnimation is null || entry.V2.MiniAnimation is null)
 		{
-			if (!DB.charAnimations.TryGetValue(entry.Configuration.Deck.Key(), out var charAnimations))
+			if (!DB.charAnimations.TryGetValue(entry.CharacterType, out var charAnimations))
 			{
 				this.LoggerProvider(entry.ModOwner).LogError($"Could not inject character {{Character}}: the `neutral` and `mini` animations are not registered.", entry.UniqueName);
 				return;
@@ -260,40 +479,81 @@ internal sealed class CharacterManager
 			}
 		}
 
-		DB.charPanels[entry.Configuration.Deck.Key()] = entry.Configuration.BorderSprite;
+		if (entry.V2.BorderSprite is { } borderSprite)
+			DB.charPanels[entry.CharacterType] = borderSprite;
+		
 		NewRunOptions.allChars = NewRunOptions.allChars
-			.Append(entry.Configuration.Deck)
+			.Append(entry.V2.Deck)
 			.Select(this.Decks.LookupByDeck)
 			.Where(e => e is not null)
 			.Select(e => e!)
 			.OrderBy(e => e.ModOwner == this.VanillaModManifest ? "" : e.ModOwner.UniqueName)
 			.Select(e => e.Deck)
 			.ToList();
-#pragma warning disable CS0618 // Type or member is obsolete
-		StarterDeck.starterSets[entry.Configuration.Deck] = entry.Configuration.Starters ?? new()
-		{
-			artifacts = entry.Configuration.StarterArtifactTypes?.Select(t => (Artifact)Activator.CreateInstance(t)!).ToList() ?? [],
-			cards = entry.Configuration.StarterCardTypes?.Select(t => (Card)Activator.CreateInstance(t)!).ToList() ?? []
-		};
-#pragma warning restore CS0618 // Type or member is obsolete
-
-		StatusMeta.deckToMissingStatus[entry.Configuration.Deck] = entry.MissingStatus.Status;
+		
+		StarterDeck.starterSets[entry.V2.Deck] = entry.V2.Starters;
+		StatusMeta.deckToMissingStatus[entry.V2.Deck] = entry.MissingStatus.Status;
 
 		this.InjectLocalization(DB.currentLocale.locale, DB.currentLocale.strings, entry);
 	}
-
-	private void InjectLocalization(string locale, Dictionary<string, string> localizations, CharacterEntry entry)
+	
+	private void Inject(NonPlayableCharacterEntry entry)
 	{
-		if (entry.Configuration.Description.Localize(locale) is { } description)
+		if (entry.V2.NeutralAnimation is { } neutralAnimationConfiguration)
 		{
-			localizations[$"char.{entry.Configuration.Deck.Key()}.desc"] = description;
-			localizations[$"char.{entry.Configuration.Deck}.desc"] = description;
+			if (neutralAnimationConfiguration.LoopTag != "neutral")
+			{
+				this.LoggerProvider(entry.ModOwner).LogError($"Could not inject character {{Character}}: `{nameof(CharacterConfiguration.NeutralAnimation)}` is not tagged `neutral`.", entry.UniqueName);
+				return;
+			}
+			this.RegisterCharacterAnimationV2(entry.ModOwner, $"{entry.UniqueName}::neutral", neutralAnimationConfiguration);
 		}
-		if (this.Decks.LookupByDeck(entry.Configuration.Deck) is { } deckEntry)
+
+		if (entry.V2.NeutralAnimation is null)
+		{
+			if (!DB.charAnimations.TryGetValue(entry.CharacterType, out var charAnimations))
+			{
+				this.LoggerProvider(entry.ModOwner).LogError($"Could not inject character {{Character}}: the `neutral` and `mini` animations are not registered.", entry.UniqueName);
+				return;
+			}
+			if (!charAnimations.ContainsKey("neutral"))
+			{
+				this.LoggerProvider(entry.ModOwner).LogError($"Could not inject character {{Character}}: the `neutral` animation is not registered.", entry.UniqueName);
+				return;
+			}
+			if (!charAnimations.ContainsKey("mini"))
+			{
+				this.LoggerProvider(entry.ModOwner).LogError($"Could not inject character {{Character}}: the `mini` animation is not registered.", entry.UniqueName);
+				return;
+			}
+		}
+
+		if (entry.V2.BorderSprite is { } borderSprite)
+			DB.charPanels[entry.CharacterType] = borderSprite;
+
+		InjectLocalization(DB.currentLocale.locale, DB.currentLocale.strings, entry);
+	}
+
+	private void InjectLocalization(string locale, Dictionary<string, string> localizations, PlayableCharacterEntry entry)
+	{
+		if (entry.V2.Description.Localize(locale) is { } description)
+		{
+			localizations[$"char.{entry.V2.Deck.Key()}.desc"] = description;
+			localizations[$"char.{entry.V2.Deck}.desc"] = description;
+		}
+		if (this.Decks.LookupByDeck(entry.V2.Deck) is { } deckEntry)
 		{
 			var characterName = deckEntry.Configuration.Name?.Invoke(locale);
-			localizations[$"char.{entry.Configuration.Deck}.desc.missing"] = $"<c={deckEntry.Configuration.Definition.color}>{characterName?.ToUpper()}..?</c>\n{characterName} is missing.";
+			localizations[$"char.{entry.V2.Deck}.desc.missing"] = $"<c={deckEntry.Configuration.Definition.color}>{characterName?.ToUpper()}..?</c>\n{characterName} is missing.";
 		}
+	}
+
+	private static void InjectLocalization(string locale, Dictionary<string, string> localizations, NonPlayableCharacterEntry entry)
+	{
+		if (entry.V2.Name.Localize(locale) is not { } name)
+			return;
+		localizations[$"char.{entry.CharacterType}"] = name;
+		localizations[$"char.{entry.CharacterType}.name"] = name;
 	}
 
 	private void OnCrystallizedFriendEvent(object? _, List<Choice> choices)
@@ -321,11 +581,13 @@ internal sealed class CharacterManager
 
 	private void OnModifyPotentialExeCards(object? _, StatePatches.ModifyPotentialExeCardsEventArgs e)
 	{
-		foreach (var character in this.UniqueNameToCharacterEntry.Values)
+		foreach (var character in this.UniqueNameToPlayableCharacterEntry.Values)
 		{
-			if (character.Configuration.ExeCardType is not { } exeCardType)
+			var configurationV2 = ((IPlayableCharacterEntryV2)character).Configuration;
+			
+			if (configurationV2.ExeCardType is not { } exeCardType)
 				continue;
-			if (e.Characters.Contains(character.Configuration.Deck))
+			if (e.Characters.Contains(configurationV2.Deck))
 				continue;
 			if (e.ExeCards.Any(c => c.GetType() == exeCardType))
 				continue;
@@ -339,9 +601,12 @@ internal sealed class CharacterManager
 			if (this.Decks.LookupByDeck(deck) is null)
 				unlockedCharacters.Remove(deck);
 
-		foreach (var entry in this.UniqueNameToCharacterEntry.Values)
-			if (!entry.Configuration.StartLocked)
-				unlockedCharacters.Add(entry.Configuration.Deck);
+		foreach (var entry in this.UniqueNameToPlayableCharacterEntry.Values)
+		{
+			var configurationV2 = ((IPlayableCharacterEntryV2)entry).Configuration;
+			if (!configurationV2.StartLocked)
+				unlockedCharacters.Add(configurationV2.Deck);
+		}
 	}
 
 	private void OnGetAssignableStatuses(object? _, WizardPatches.GetAssignableStatusesEventArgs e)
@@ -351,26 +616,65 @@ internal sealed class CharacterManager
 		{
 			if (character.deckType is not { } deck)
 				continue;
-			if (this.UniqueNameToCharacterEntry.Values.FirstOrDefault(e => e.Configuration.Deck == deck) is not { } entry)
+			if (this.UniqueNameToPlayableCharacterEntry.Values.FirstOrDefault(e => e.V2.Deck == deck) is not { } entry)
 				continue;
 			e.Statuses.Add(entry.MissingStatus.Status);
 		}
 	}
 
-	private sealed class AnimationEntry(IModManifest modOwner, string uniqueName, CharacterAnimationConfiguration configuration)
-		: ICharacterAnimationEntry
+	private sealed class AnimationEntry(
+		IModManifest modOwner,
+		string uniqueName,
+		CharacterAnimationConfiguration? v1,
+		CharacterAnimationConfigurationV2 v2
+	) : ICharacterAnimationEntry, ICharacterAnimationEntryV2
 	{
+		internal CharacterAnimationConfiguration? V1 { get; } = v1;
+		internal CharacterAnimationConfigurationV2 V2 { get; } = v2;
+		
 		public IModManifest ModOwner { get; } = modOwner;
 		public string UniqueName { get; } = uniqueName;
-		public CharacterAnimationConfiguration Configuration { get; } = configuration;
+
+		CharacterAnimationConfiguration ICharacterAnimationEntry.Configuration => this.V1 ?? throw new InvalidOperationException();
+		
+		CharacterAnimationConfigurationV2 ICharacterAnimationEntryV2.Configuration => this.V2;
 	}
 
-	private sealed class CharacterEntry(IModManifest modOwner, string uniqueName, CharacterConfiguration configuration, IStatusEntry missingStatus)
-		: ICharacterEntry
+	private sealed class PlayableCharacterEntry(
+		IModManifest modOwner,
+		string uniqueName,
+		CharacterConfiguration v1,
+		PlayableCharacterConfigurationV2 v2,
+		IStatusEntry missingStatus
+	) : ICharacterEntry, IPlayableCharacterEntryV2
 	{
+		internal CharacterConfiguration V1 { get; } = v1;
+		internal PlayableCharacterConfigurationV2 V2 { get; } = v2;
+		
 		public IModManifest ModOwner { get; } = modOwner;
 		public string UniqueName { get; } = uniqueName;
-		public CharacterConfiguration Configuration { get; } = configuration;
 		public IStatusEntry MissingStatus { get; } = missingStatus;
+		public string CharacterType => this.V2.Deck == Deck.colorless ? "comp" : this.V2.Deck.Key();
+
+		CharacterConfiguration ICharacterEntry.Configuration => this.V1;
+		
+		PlayableCharacterConfigurationV2 IPlayableCharacterEntryV2.Configuration => this.V2;
+		Spr? ICharacterEntryV2.BorderSprite => this.V2.BorderSprite;
+	}
+
+	private sealed class NonPlayableCharacterEntry(
+		IModManifest modOwner,
+		string uniqueName,
+		NonPlayableCharacterConfigurationV2 v2
+	) : INonPlayableCharacterEntryV2
+	{
+		internal NonPlayableCharacterConfigurationV2 V2 { get; } = v2;
+		
+		public IModManifest ModOwner { get; } = modOwner;
+		public string UniqueName { get; } = uniqueName;
+		public string CharacterType => this.V2.CharacterType;
+		
+		NonPlayableCharacterConfigurationV2 INonPlayableCharacterEntryV2.Configuration => this.V2;
+		Spr? ICharacterEntryV2.BorderSprite => this.V2.BorderSprite;
 	}
 }