+++ Nickel/Interfaces/Content/ICharacterAnimationEntryV2.cs
@@ -1,0 +1,10 @@
+namespace Nickel;
+
+/// <summary>
+/// Describes an animation for a <see cref="Character"/>.
+/// </summary>
+public interface ICharacterAnimationEntryV2 : IModOwned
+{
+	/// <summary>The configuration used to register the <see cref="Character"/> animation.</summary>
+	CharacterAnimationConfigurationV2 Configuration { get; }
+}