+++ docs/release-notes.md
@@ -20,6 +20,7 @@
 
 ### For developers:
 * Added APIs for storing mod settings in a common place.
+* Added APIs for registering non-playable characters (enemies or story characters).
 * Added an event mods can subscribe to to get informed about the game closing, either normally or via an exception being thrown.
 * Added a way for mods to get a list of all resolved mods (including ones which may not have been loaded, for various reasons).
 * Added additional APIs for obtaining new collision-free enum case values.